Refactor code to use long instead of int for originalLength in DefiniteLengthInputStream and update method signatures in Streams.cs
This commit changes the type of _originalLength from int to long in DefiniteLengthInputStream. It also updates the method signatures in Streams.cs to reflect this change, replacing all instances of "long limit" with "int limit".

+++ src/core/IO/DefiniteLengthInputStream.cs
@@ -22,7 +22,7 @@ public class DefiniteLengthInputStream : LimitedInputStream
 	/// <summary>
 	///   The original length
 	/// </summary>
-	private readonly int _originalLength;
+	private readonly long _originalLength;
 
 	/// <summary>
 	///   Initializes a new instance of the <see cref="DefiniteLengthInputStream" /> class.
@@ -114,8 +114,7 @@ public class DefiniteLengthInputStream : LimitedInputStream
 	/// <exception cref="System.IO.EndOfStreamException">DEF length " + _originalLength + " object truncated by " + Remaining</exception>
 	public override int Read(Span<byte> buffer)
 	{
-		if (Remaining == 0)
-			return 0;
+		if (Remaining == 0) return 0;
 
 		var toRead  = System.Math.Min(buffer.Length, Remaining);
 		var numRead = Stream.Read(buffer[..toRead]);
